# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Groene Vingers",
        "version": "3.0.0"
    },
    "paths": {
...
        "/api_gv/public/api/customer": {
            "post": {
                "tags": [
                    "Customer Order"
                ],
                "summary": "Customer Order",
                "description": "Customer Order",
                "operationId": "store_customer_order",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Hasan"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0681302145"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "hasan@gmail.com"
                                    },
                                    "zipcode": {
                                        "type": "string",
                                        "example": "5737GB"
                                    },
                                    "huisnummer": {
                                        "type": "string",
                                        "example": "71"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Dorpsstraat 71, 5737GB Lieshout"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Lieshout"
                                    },
                                    "total": {
                                        "type": "integer",
                                        "example": "200"
                                    },
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Product 50"
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": "4"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Customer Order Created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api_gv/public/api/kuin/products": {
            "get": {
                "tags": [
                    "Product Kuin"
                ],
                "summary": "Get product List",
                "description": "Get product List as Array",
                "operationId": "index_kuin",
                "responses": {
                    "200": {
                        "description": "Get product List as Array"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api_gv/public/api/user/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Show user info",
                "description": "Show user info",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show user info"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user",
                "description": "Update user",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "name"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "lastname"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "city"
                                    },
                                    "phone": {
                                        "type": "integer",
                                        "example": "000000"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "manager@groenevingers.com"
                                    },
                                    "birthday": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "17-08-1990"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "manager"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update user"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete user member",
                "description": "Delete user member",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete user member"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    ]
}